Move QueriesPage to admin
