--- ui/src/queries/index.js
@@ -1,2 +1,0 @@
-import QueriesPage from './containers/QueriesPage'
-export default QueriesPage
